Split out Unicode matching, make filters more explicit.
This makes the filters in "text-fuzzy.c.in" more explicit, adds some
new filters, adds a new field giving the number of things rejected by
length, and puts more messages in.

There is a new function "length_rejections" which returns the number
of elements in "nearest" which were rejected by the length filter.

Test "t/private-functions.t" for the private functions documented.