Mistake about max/min in trans; memory for Solaris
Some compiler failures like this on non-GCC compilers:

http://www.cpantesters.org/cpan/report/b2ef23ee-c749-11e8-b938-de915b3e2777

These were caused by CALLOC macro with sizeof left in.

Also, there was a mistake in the max limiting code where it should
look at the minimum on a row and reject the match if that is over the
max, not the other way around. Because we only need the minimum, I
removed unneeded "maximum" routine from text-fuzzy.c.tmpl.

Very minimal test about the mistake.